# api/openapi.yaml
openapi: 3.1.0
info:
  title: SmartCare PRO API
  description: |
    Comprehensive Electronic Health Record (EHR) system for Zambian healthcare,
    focusing on intelligent patient transfer and clinical decision support.
    
    ## Features
    - WHO guideline-compliant diagnostic protocols
    - Advanced clinical intelligence with Anthropic AI
    - Smart Inter-Facility Patient Transfer system
    - Comprehensive patient registry (185 columns)
    - ANC clinical decision support (83 columns)
    - Real-time clinical alerts and recommendations
    
    ## Authentication
    Uses Replit Auth integration with session-based authentication.
    All endpoints require valid authentication except for health checks.
  version: 1.3.0
  contact:
    name: Institute for Health Measurement (IHM) Southern Africa
    url: https://smartcare.gov.zm
  license:
    name: Ministry of Health - Zambia
    url: https://moh.gov.zm
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://smartcare.replit.app
    description: Production server
security:
  - cookieAuth: []
paths:
...
  /api/admin/cache/stats:
    get:
      summary: Get cache statistics and health
      tags: [Cache Management]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Cache statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      totalKeys:
                        type: integer
                      memoryUsage:
                        type: string
                      hitRate:
                        type: number
        '403':
          description: Admin access required
        '500':
          description: Server error
  /api/admin/cache/clear:
    delete:
      summary: Clear all cached data
      tags: [Cache Management]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Cache cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '403':
          description: Admin access required
        '500':
          description: Server error
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie authentication
...